Item pool updates
