+++ src/Randomizer.Data/WorldData/Item.cs
@@ -225,260 +225,6 @@ namespace Randomizer.Data.WorldData
             return Metadata.TryGetTrackingResponse(State.TrackingState, out response);
         }
 
-
-        /// <summary>
-        /// Returns a list of the items required to progress through the game.
-        /// </summary>
-        /// <param name="world">The world to assign to the items.</param>
-        /// <returns>A new collection of items.</returns>
-        public static List<Item> CreateProgressionPool(World world)
-        {
-            var itemPool = new List<Item> {
-                new Item(ItemType.ProgressiveShield, world),
-                new Item(ItemType.ProgressiveShield, world),
-                new Item(ItemType.ProgressiveShield, world),
-                new Item(ItemType.ProgressiveSword, world),
-                new Item(ItemType.ProgressiveSword, world),
-                new Item(ItemType.Bow, world),
-                new Item(ItemType.Hookshot, world),
-                new Item(ItemType.Mushroom, world),
-                new Item(ItemType.Powder, world),
-                new Item(ItemType.Firerod, world),
-                new Item(ItemType.Icerod, world),
-                new Item(ItemType.Bombos, world),
-                new Item(ItemType.Ether, world),
-                new Item(ItemType.Quake, world),
-                new Item(ItemType.Lamp, world),
-                new Item(ItemType.Hammer, world),
-                new Item(ItemType.Shovel, world),
-                new Item(ItemType.Flute, world),
-                new Item(ItemType.Book, world),
-                new Item(ItemType.Bottle, world),
-                new Item(ItemType.Somaria, world),
-                new Item(ItemType.Byrna, world),
-                new Item(ItemType.Cape, world),
-                new Item(ItemType.Mirror, world),
-                new Item(ItemType.Boots, world),
-                new Item(ItemType.ProgressiveGlove, world),
-                new Item(ItemType.ProgressiveGlove, world),
-                new Item(ItemType.Flippers, world),
-                new Item(ItemType.MoonPearl, world),
-                new Item(ItemType.HalfMagic, world),
-
-                new Item(ItemType.Grapple, world),
-                new Item(ItemType.Charge, world),
-                new Item(ItemType.Ice, world),
-                new Item(ItemType.Wave, world),
-                new Item(ItemType.Plasma, world),
-                new Item(ItemType.Varia, world),
-                new Item(ItemType.Gravity, world),
-                new Item(ItemType.Morph, world),
-                new Item(ItemType.Bombs, world),
-                new Item(ItemType.SpringBall, world),
-                new Item(ItemType.ScrewAttack, world),
-                new Item(ItemType.HiJump, world),
-                new Item(ItemType.SpaceJump, world),
-                new Item(ItemType.SpeedBooster, world),
-
-                new Item(ItemType.Missile, world),
-                new Item(ItemType.Super, world),
-                new Item(ItemType.PowerBomb, world),
-                new Item(ItemType.PowerBomb, world),
-                new Item(ItemType.ETank, world),
-                new Item(ItemType.ETank, world),
-                new Item(ItemType.ETank, world),
-                new Item(ItemType.ETank, world),
-                new Item(ItemType.ETank, world),
-
-                new Item(ItemType.ReserveTank, world),
-                new Item(ItemType.ReserveTank, world),
-                new Item(ItemType.ReserveTank, world),
-                new Item(ItemType.ReserveTank, world),
-
-                new Item(ItemType.ThreeHundredRupees, world),
-                new Item(ItemType.ThreeHundredRupees, world),
-            };
-
-            foreach (var item in itemPool)
-            {
-                item.Progression = true;
-            }
-
-            return itemPool;
-        }
-
-        /// <summary>
-        /// Returns a list of items that are nice to have but are not required
-        /// to finish the game.
-        /// </summary>
-        /// <param name="world">The world to assign to the items.</param>
-        /// <returns>A new collection of items.</returns>
-        public static List<Item> CreateNicePool(World world)
-        {
-            var itemPool = new List<Item> {
-                new Item(ItemType.ProgressiveTunic, world),
-                new Item(ItemType.ProgressiveTunic, world),
-                new Item(ItemType.ProgressiveSword, world),
-                new Item(ItemType.ProgressiveSword, world),
-                new Item(ItemType.SilverArrows, world),
-                new Item(ItemType.BlueBoomerang, world),
-                new Item(ItemType.RedBoomerang, world),
-                new Item(ItemType.Bottle, world),
-                new Item(ItemType.Bottle, world),
-                new Item(ItemType.Bottle, world),
-                new Item(ItemType.Bugnet, world),
-                new Item(ItemType.HeartContainerRefill, world),
-
-                new Item(ItemType.Spazer, world),
-                new Item(ItemType.XRay, world),
-            };
-
-            if (world.Config.CasPatches.ExpandedItems)
-            {
-                itemPool.Add(new Item(ItemType.HalfMagic, world));
-            }
-
-            itemPool.AddRange(Copies(10, () => new Item(ItemType.HeartContainer, world)));
-
-            return itemPool;
-        }
-
-        /// <summary>
-        /// Returns a list of items used to fill remaining locations.
-        /// </summary>
-        /// <param name="world">The world to assign to the items.</param>
-        /// <returns>A new collection of items.</returns>
-        public static List<Item> CreateJunkPool(World world)
-        {
-            var itemPool = new List<Item> {
-                new Item(ItemType.Arrow, world),
-                new Item(ItemType.OneHundredRupees, world)
-            };
-
-            itemPool.AddRange(Copies(24, () => new Item(ItemType.HeartPiece, world)));
-            itemPool.AddRange(Copies(8, () => new Item(ItemType.TenArrows, world)));
-            itemPool.AddRange(Copies(13, () => new Item(ItemType.ThreeBombs, world)));
-            itemPool.AddRange(Copies(4, () => new Item(ItemType.ArrowUpgrade5, world)));
-            itemPool.AddRange(Copies(4, () => new Item(ItemType.BombUpgrade5, world)));
-            itemPool.AddRange(Copies(2, () => new Item(ItemType.OneRupee, world)));
-            itemPool.AddRange(Copies(4, () => new Item(ItemType.FiveRupees, world)));
-            itemPool.AddRange(Copies(world.Config.MetroidKeysanity ? 24 : 27, () => new Item(ItemType.TwentyRupees, world)));
-            itemPool.AddRange(Copies(7, () => new Item(ItemType.FiftyRupees, world)));
-            itemPool.AddRange(Copies(3, () => new Item(ItemType.ThreeHundredRupees, world)));
-
-            itemPool.AddRange(Copies(9, () => new Item(ItemType.ETank, world)));
-            itemPool.AddRange(Copies(39, () => new Item(ItemType.Missile, world)));
-            itemPool.AddRange(Copies(15, () => new Item(ItemType.Super, world)));
-            itemPool.AddRange(Copies(8, () => new Item(ItemType.PowerBomb, world)));
-
-            // If no quarter magic, add an extra 20 rupees back
-            if (!world.Config.CasPatches.ExpandedItems)
-            {
-                itemPool.Add(new Item(ItemType.TwentyRupees, world));
-            }
-
-            return itemPool;
-        }
-
-        /// <summary>
-        /// Returns a list of dungeon-specific items.
-        /// </summary>
-        /// <remarks>The order of the dungeon pool is significant.</remarks>
-        /// <param name="world">The world to assign to the items.</param>
-        /// <returns>A new collection of items.</returns>
-        public static List<Item> CreateDungeonPool(World world)
-        {
-            var itemPool = new List<Item>();
-
-            itemPool.AddRange(new[] {
-                new Item(ItemType.BigKeyEP, world),
-                new Item(ItemType.BigKeyDP, world),
-                new Item(ItemType.BigKeyTH, world),
-                new Item(ItemType.BigKeyPD, world),
-                new Item(ItemType.BigKeySP, world),
-                new Item(ItemType.BigKeySW, world),
-                new Item(ItemType.BigKeyTT, world),
-                new Item(ItemType.BigKeyIP, world),
-                new Item(ItemType.BigKeyMM, world),
-                new Item(ItemType.BigKeyTR, world),
-                new Item(ItemType.BigKeyGT, world),
-            });
-
-            itemPool.AddRange(Copies(1, () => new Item(ItemType.KeyHC, world)));
-            itemPool.AddRange(Copies(2, () => new Item(ItemType.KeyCT, world)));
-            itemPool.AddRange(Copies(1, () => new Item(ItemType.KeyDP, world)));
-            itemPool.AddRange(Copies(1, () => new Item(ItemType.KeyTH, world)));
-            itemPool.AddRange(Copies(6, () => new Item(ItemType.KeyPD, world)));
-            itemPool.AddRange(Copies(1, () => new Item(ItemType.KeySP, world)));
-            itemPool.AddRange(Copies(3, () => new Item(ItemType.KeySW, world)));
-            itemPool.AddRange(Copies(1, () => new Item(ItemType.KeyTT, world)));
-            itemPool.AddRange(Copies(2, () => new Item(ItemType.KeyIP, world)));
-            itemPool.AddRange(Copies(3, () => new Item(ItemType.KeyMM, world)));
-            itemPool.AddRange(Copies(4, () => new Item(ItemType.KeyTR, world)));
-            itemPool.AddRange(Copies(4, () => new Item(ItemType.KeyGT, world)));
-
-            itemPool.AddRange(new[] {
-                new Item(ItemType.MapEP, world),
-                new Item(ItemType.MapDP, world),
-                new Item(ItemType.MapTH, world),
-                new Item(ItemType.MapPD, world),
-                new Item(ItemType.MapSP, world),
-                new Item(ItemType.MapSW, world),
-                new Item(ItemType.MapTT, world),
-                new Item(ItemType.MapIP, world),
-                new Item(ItemType.MapMM, world),
-                new Item(ItemType.MapTR, world),
-            });
-            if (!world.Config.MetroidKeysanity)
-            {
-                itemPool.AddRange(new[] {
-                    new Item(ItemType.MapHC, world),
-                    new Item(ItemType.MapGT, world),
-                    new Item(ItemType.CompassEP, world),
-                    new Item(ItemType.CompassDP, world),
-                    new Item(ItemType.CompassTH, world),
-                    new Item(ItemType.CompassPD, world),
-                    new Item(ItemType.CompassSP, world),
-                    new Item(ItemType.CompassSW, world),
-                    new Item(ItemType.CompassTT, world),
-                    new Item(ItemType.CompassIP, world),
-                    new Item(ItemType.CompassMM, world),
-                    new Item(ItemType.CompassTR, world),
-                    new Item(ItemType.CompassGT, world),
-                });
-            }
-
-            return itemPool;
-        }
-
-        /// <summary>
-        /// Returns a list of the items required to open doors in Metroid in a
-        /// keysanity seed.
-        /// </summary>
-        /// <param name="world">The world to assign to the items.</param>
-        /// <returns>A new collection of items.</returns>
-        public static List<Item> CreateKeycards(World world)
-        {
-            return new List<Item> {
-                new Item(ItemType.CardCrateriaL1, world),
-                new Item(ItemType.CardCrateriaL2, world),
-                new Item(ItemType.CardCrateriaBoss, world),
-                new Item(ItemType.CardBrinstarL1, world),
-                new Item(ItemType.CardBrinstarL2, world),
-                new Item(ItemType.CardBrinstarBoss, world),
-                new Item(ItemType.CardNorfairL1, world),
-                new Item(ItemType.CardNorfairL2, world),
-                new Item(ItemType.CardNorfairBoss, world),
-                new Item(ItemType.CardMaridiaL1, world),
-                new Item(ItemType.CardMaridiaL2, world),
-                new Item(ItemType.CardMaridiaBoss, world),
-                new Item(ItemType.CardWreckedShipL1, world),
-                new Item(ItemType.CardWreckedShipBoss, world),
-                new Item(ItemType.CardLowerNorfairL1, world),
-                new Item(ItemType.CardLowerNorfairBoss, world),
-            };
-        }
-
         /// <summary>
         /// Determines whether the item is of the specified type and belongs to
         /// the specified world.
