+++ tests/Randomizer.SMZ3.Tests/LogicTests/LocationLogicTests.cs
@@ -61,7 +61,7 @@ namespace Randomizer.SMZ3.Tests.LogicTests
         [Fact]
         public void LocationWithTwoMissingItemsReturnsTwoMissingItems()
         {
-            var emptyProgression = new Progression(Item.CreateKeycards(World), new List<Reward>(), new List<Boss>());
+            var emptyProgression = new Progression(World.ItemPools.Keycards, new List<Reward>(), new List<Boss>());
             var missingItems = Logic.GetMissingRequiredItems(World.GreenBrinstar.PowerBomb, emptyProgression, out _);
             missingItems.Should().ContainEquivalentOf(new[] { ItemType.Morph, ItemType.PowerBomb });
         }
@@ -69,7 +69,7 @@ namespace Randomizer.SMZ3.Tests.LogicTests
         [Fact]
         public void LocationWithMultipleOptionsReturnsAllOptions()
         {
-            var emptyProgression = new Progression(Item.CreateKeycards(World), new List<Reward>(), new List<Boss>());
+            var emptyProgression = new Progression(World.ItemPools.Keycards, new List<Reward>(), new List<Boss>());
             var missingItems = Logic.GetMissingRequiredItems(World.BlueBrinstar.Ceiling, emptyProgression, out _);
             missingItems.Should().ContainEquivalentOf(new[] { ItemType.SpaceJump })
                 .And.ContainEquivalentOf(new[] { ItemType.HiJump })
@@ -80,7 +80,7 @@ namespace Randomizer.SMZ3.Tests.LogicTests
         [Fact]
         public void LocationWithThreeMissingItemsReturnsThreeMissingItems()
         {
-            var emptyProgression = new Progression(Item.CreateKeycards(World), new List<Reward>(), new List<Boss>());
+            var emptyProgression = new Progression(World.ItemPools.Keycards, new List<Reward>(), new List<Boss>());
             var missingItems = Logic.GetMissingRequiredItems(World.CentralCrateria.BombTorizo, emptyProgression, out _);
             missingItems.Should().ContainEquivalentOf(new[] { ItemType.Morph, ItemType.Super, ItemType.PowerBomb });
         }
